Fixed the cooldowns not showing on items when used. Fixed the items not greying themselves when they are put into mail as attachments
Both fixes have to call updateBagSize for each bag *and then* updateItem for each item for each bag, in order to work.
I'm not sure why this is as I'm new to the code

updateItemCooldowndoesn't work as item[1] is always nil... This function needs looking at if it's crucial that it works.

Had a wild stab at trying to cure the shuffle of the bottom half of the bags (in the most inefficient and horrible way ever) by adding ZONE_CHANGED_NEW_AREA event listener and updating each bag's type and size.
This isn't confirmed to work at the moment, but I've yet to see a shuffle with this in place.
If this does work it doesn't solve the root of the problem, which Carsight exaplins on curse in post: http://www.curse.com/addons/wow/combuctor?comment=543